Convert parameter target utils to TypeScript (#24571)
