+++ frontend/src/metabase/parameters/utils/targets.ts
@@ -1,0 +1,43 @@
+import Dimension from "metabase-lib/lib/Dimension";
+import {
+  ParameterTarget,
+  ParameterDimensionTarget,
+} from "metabase-types/types/Parameter";
+import Metadata from "metabase-lib/lib/metadata/Metadata";
+import Question from "metabase-lib/lib/Question";
+import StructuredQuery from "metabase-lib/lib/queries/StructuredQuery";
+import NativeQuery from "metabase-lib/lib/queries/NativeQuery";
+
+export function isDimensionTarget(
+  target: ParameterTarget,
+): target is ParameterDimensionTarget {
+  return target?.[0] === "dimension";
+}
+
+export function isVariableTarget(target: ParameterTarget) {
+  return target?.[0] === "variable";
+}
+
+export function getTemplateTagFromTarget(target: ParameterTarget) {
+  if (!target?.[1]) {
+    return null;
+  }
+
+  const [, [type, tag]] = target;
+  return type === "template-tag" ? tag : null;
+}
+
+export function getParameterTargetField(
+  target: ParameterTarget,
+  metadata: Metadata,
+  question: Question,
+) {
+  if (isDimensionTarget(target)) {
+    const query = question.query() as NativeQuery | StructuredQuery;
+    const dimension = Dimension.parseMBQL(target[1], metadata, query);
+
+    return dimension?.field();
+  }
+
+  return null;
+}
